move tests to different suite
